Use coordinate*10^7 as int instead of double

+++ src/graphrep/NNSearcher.java
@@ -7,5 +7,5 @@ public interface NNSearcher extends Serializable {
 	// it's best to use the protected lon[] and lat[] arrays in GraphRep for
 	// accessing all nodes
 
-	int getIDForCoordinates(double lat, double lon);
+	int getIDForCoordinates(int lat, int lon);
 }
